Logic moved to RequestService, PUT does not work yet
